pyproject.toml
[project]
name = "rag-skills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "loguru>=0.7.3",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pypdf>=6.9.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.42.0",
]
